# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DailyScoreInput {
  /**
     * @minLength 1
     * @maxLength 24
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 36000
     */
  seconds: number;
}

export interface DailyScoreEntry {
  name: string;
  seconds: number;
}

export interface DailyLeaderboardResponse {
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  date: string;
  entries: DailyScoreEntry[];
}

export interface ErrorResponse {
  error: string;
}