# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: daily
    description: Shared daily puzzle leaderboard operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /daily/leaderboard:
    get:
      operationId: getDailyLeaderboard
      tags: [daily]
      summary: Get today's shared leaderboard
      description: Returns the top daily puzzle times for the current UTC date.
      responses:
        "200":
          description: Today's leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLeaderboardResponse"
    post:
      operationId: submitDailyScore
      tags: [daily]
      summary: Submit a daily puzzle time
      description: Adds a completed daily time to the shared leaderboard for the current UTC date.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyScoreInput"
      responses:
        "200":
          description: Updated leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLeaderboardResponse"
        "400":
          description: Invalid score submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    DailyScoreInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 24
        seconds:
          type: number
          minimum: 1
          maximum: 36000
      required:
        - name
        - seconds
    DailyScoreEntry:
      type: object
      properties:
        name:
          type: string
        seconds:
          type: number
      required:
        - name
        - seconds
    DailyLeaderboardResponse:
      type: object
      properties:
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
        entries:
          type: array
          items:
            $ref: "#/components/schemas/DailyScoreEntry"
      required:
        - date
        - entries
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error