# lib/api-zod/src/generated/types/dailyLeaderboardResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DailyScoreEntry } from './dailyScoreEntry';

export interface DailyLeaderboardResponse {
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  date: string;
  entries: DailyScoreEntry[];
}