# lib/api-zod/src/generated/types/dailyScoreInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DailyScoreInput {
  /**
     * @minLength 1
     * @maxLength 24
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 36000
     */
  seconds: number;
}